feat: use kdtree when weights are zeros in mesh diagram fn
