+++ src/synthetmic/_internal/_errors.py
@@ -1,0 +1,4 @@
+class _NotFittedError(ValueError, AttributeError):
+    """
+    Raised when attempting to use an unfitted generator.
+    """
